requirements.txt
flask
requests
google-cloud-translate
google-cloud-texttospeech
# Add any other necessary libraries